refactor: eqipment folder to equipments
