+++ src/main/kotlin/Application.kt
@@ -5,7 +5,7 @@ import at.eventful.messless.plugins.db.configureDatabases
 import at.eventful.messless.plugins.socket.WebSocketRouter
 import at.eventful.messless.plugins.socket.configureWebSocket
 import at.eventful.messless.services.echo.EchoService
-import at.eventful.messless.services.eqipments.EquipmentsService
+import at.eventful.messless.services.equipments.EquipmentsService
 import at.eventful.messless.services.events.EventsService
 import at.eventful.messless.services.index.registerIndexRoute
 import at.eventful.messless.services.technicalLogEntries.TechnicalLogEntriesService
